+++ HISTORY.rst
@@ -18,6 +18,7 @@ History
     * Fix: validate column filters before trying to access attr to create the filter (rudaporto).
     * Improve base resource filters to be filter association proxy attributes using special comparator (rudaporto).
     * Validate if we can find the column to do the filer and send the proper error messsage if not (rudaporto).
+    * Force ObjectCreatedEvent and ObjectUpdateEvent to call BaseEvent.__call__ method (rudaporto).
 
 
 2.0.6 (2017-07-27)